# back/api.yaml
openapi: 3.0.0
paths:
...
  /feature-flag:
    get:
      operationId: FeatureFlagController_search
      parameters:
        - name: page
          required: false
          in: query
          description: The page number
          example: 1
          schema:
            type: number
        - name: perPage
          required: false
          in: query
          description: The number of items per page
          example: 10
          schema:
            type: number
        - name: sort
          required: false
          in: query
          description: The field that should be used for sorting
          example: name
          schema:
            nullable: true
            enum:
              - name
              - createdAt
            type: string
        - name: sortDir
          required: false
          in: query
          description: The sort direction
          example: asc
          schema:
            nullable: true
            enum:
              - asc
              - desc
            type: string
        - name: filter
          required: false
          in: query
          description: The filter to apply to the search
          example:
            name: feature
            description: user feature
            enabled: true
            userId: uuid
          schema:
            nullable: true
      responses:
        '200':
          description: The FeatureFlag has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagCollectionPresenter'
        '401':
          description: Unauthorized
        '422':
          description: Unprocessable Entity
      tags:
        - feature-flag
    post:
      operationId: FeatureFlagController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureFlagDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagPresenter'
        '401':
          description: Unauthorized
        '422':
          description: Unprocessable Entity
      tags:
        - feature-flag
      security:
        - bearer: []
  /feature-flag/id:
    get:
      operationId: FeatureFlagController_findManyById
      parameters:
        - name: id
          required: true
          in: query
          description: >-
            Feature flag identifiers to be retrieved. Accepts comma-separated or
            repeated query params.
          example:
            - 8cf10340-8c08-4f2d-8c88-ef1d8f8536da
          schema:
            type: array
            items:
              type: array
              items:
                type: a
        - name: appUserId
          required: false
          in: query
          description: >-
            Optional AppUser identifier to filter feature flags that target the
            informed user.
          example: 5f2c3de7-0fc9-475c-a6b4-d9067bda3d89
          schema:
            type: string
      responses:
        '200':
          description: FeatureFlags fetched successfully by identifiers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagCollectionPresenter'
        '401':
          description: Unauthorized
      tags:
        - feature-flag
      security:
        - bearer: []
  /feature-flag/{id}:
    get:
      operationId: FeatureFlagController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagPresenter'
        '401':
          description: Unauthorized
        '404':
          description: FeatureFlag not found
      tags:
        - feature-flag
      security:
        - bearer: []
    put:
      operationId: FeatureFlagController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFeatureFlagDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagPresenter'
        '401':
          description: Unauthorized
        '404':
          description: FeatureFlag not found
        '422':
          description: Unprocessable Entity
      tags:
        - feature-flag
      security:
        - bearer: []
    delete:
      operationId: FeatureFlagController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: FeatureFlag deleted
        '401':
          description: Unauthorized
        '404':
          description: FeatureFlag not found
      tags:
        - feature-flag
      security:
        - bearer: []
  /feature-flag/client:
    get:
      operationId: FeatureFlagController_clientFeatureFlags
      parameters:
        - name: token
          required: true
          in: query
          description: The feature flag admin api token
          example: 79fcd962-d523-4531-8b74-2854f579e062
          schema:
            type: string
        - name: featureFlagId
          required: true
          in: query
          description: The feature flag ID
          example: 9de0d313-2302-48e6-b775-6e93922ad38a
          schema:
            type: string
        - name: appUserId
          required: false
          in: query
          description: The app user id for feature flag
          example: 1953870b-41c0-4f6b-b13d-9a2c4e206d9c
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagPresenter'
        '401':
          description: Unauthorized
        '404':
          description: FeatureFlag not found
      tags:
        - feature-flag
...
info:
  title: Nest Auth Api
  description: The Nest Api support creating deleting editing and authenticating users
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: Bearer
      bearerFormat: JWT
      name: Authorization
      description: JWT token
      in: Header
      type: http
  schemas:
...
    FeatureFlagPresenter:
      type: object
      properties:
        id:
          type: string
          description: The id of the feature flag
        name:
          type: string
          description: The name of the feature flag
        description:
          type: string
          description: The description of the feature flag
        enabled:
          type: boolean
          description: The status of the feature flag, if is active or not
        createdAt:
          format: date-time
          type: string
          description: The date when the feature flag was created
        updatedAt:
          format: date-time
          type: string
          nullable: true
          description: The data of last feature flag update
      required:
        - id
        - name
        - description
        - enabled
        - createdAt
        - updatedAt
    FeatureFlagCollectionPresenter:
      type: object
      properties:
        data:
          description: List of FeatureFlags
          type: array
          items:
            $ref: '#/components/schemas/FeatureFlagPresenter'
      required:
        - data
    UpdateFeatureFlagDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the feature flag
          example: new-feature-flag
        description:
          type: string
          description: The description of the feature flag
          example: This flag enables the new user dashboard
        enabled:
          type: boolean
          description: Whether the feature flag is enabled
          example: true
      required:
        - name
        - description
        - enabled
    CreateFeatureFlagDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the feature flag
          example: new-feature-flag
        description:
          type: string
          description: The description of the feature flag
          example: This flag enables the new user dashboard
        enabled:
          type: boolean
          description: Whether the feature flag is enabled
          example: true
      required:
        - name
        - description
        - enabled
...